+++ collector/src/main/java/com/bitmovin/analytics/BitmovinAnalyticsConfig.java
@@ -16,6 +16,25 @@ public class BitmovinAnalyticsConfig implements Parcelable {
     private String customData5;
     private String customData6;
     private String customData7;
+    private String customData8;
+    private String customData9;
+    private String customData10;
+    private String customData11;
+    private String customData12;
+    private String customData13;
+    private String customData14;
+    private String customData15;
+    private String customData16;
+    private String customData17;
+    private String customData18;
+    private String customData19;
+    private String customData20;
+    private String customData21;
+    private String customData22;
+    private String customData23;
+    private String customData24;
+    private String customData25;
+
     private String customUserId;
     private String experimentName;
     private String mpdUrl;
@@ -77,6 +96,24 @@ public class BitmovinAnalyticsConfig implements Parcelable {
         customData5 = in.readString();
         customData6 = in.readString();
         customData7 = in.readString();
+        customData8 = in.readString();
+        customData9 = in.readString();
+        customData10 = in.readString();
+        customData11 = in.readString();
+        customData12 = in.readString();
+        customData13 = in.readString();
+        customData14 = in.readString();
+        customData15 = in.readString();
+        customData16 = in.readString();
+        customData17 = in.readString();
+        customData18 = in.readString();
+        customData19 = in.readString();
+        customData20 = in.readString();
+        customData21 = in.readString();
+        customData22 = in.readString();
+        customData23 = in.readString();
+        customData24 = in.readString();
+        customData25 = in.readString();
         customUserId = in.readString();
         experimentName = in.readString();
         mpdUrl = in.readString();
@@ -106,6 +143,24 @@ public class BitmovinAnalyticsConfig implements Parcelable {
         dest.writeString(customData5);
         dest.writeString(customData6);
         dest.writeString(customData7);
+        dest.writeString(customData8);
+        dest.writeString(customData9);
+        dest.writeString(customData10);
+        dest.writeString(customData11);
+        dest.writeString(customData12);
+        dest.writeString(customData13);
+        dest.writeString(customData14);
+        dest.writeString(customData15);
+        dest.writeString(customData16);
+        dest.writeString(customData17);
+        dest.writeString(customData18);
+        dest.writeString(customData19);
+        dest.writeString(customData20);
+        dest.writeString(customData21);
+        dest.writeString(customData22);
+        dest.writeString(customData23);
+        dest.writeString(customData24);
+        dest.writeString(customData25);
         dest.writeString(customUserId);
         dest.writeString(experimentName);
         dest.writeString(mpdUrl);
@@ -264,6 +319,240 @@ public class BitmovinAnalyticsConfig implements Parcelable {
      */
     public void setCustomData7(String customData7) {
         this.customData7 = customData7;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData8
+     */
+    public void setCustomData8(String customData8) {
+        this.customData8 = customData8;
+    }
+
+    public String getCustomData8() {
+        return customData8;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData9
+     */
+    public void setCustomData9(String customData9) {
+        this.customData9 = customData9;
+    }
+
+    public String getCustomData9() {
+        return customData9;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData10
+     */
+    public void setCustomData10(String customData10) {
+        this.customData10 = customData10;
+    }
+
+    public String getCustomData10() {
+        return customData10;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData11
+     */
+    public void setCustomData11(String customData11) {
+        this.customData11 = customData11;
+    }
+
+    public String getCustomData11() {
+        return customData11;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData12
+     */
+    public void setCustomData12(String customData12) {
+        this.customData12 = customData12;
+    }
+
+    public String getCustomData12() {
+        return customData12;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData13
+     */
+    public void setCustomData13(String customData13) {
+        this.customData13 = customData13;
+    }
+
+    public String getCustomData13() {
+        return customData13;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData14
+     */
+    public void setCustomData14(String customData14) {
+        this.customData14 = customData14;
+    }
+
+    public String getCustomData14() {
+        return customData14;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData15
+     */
+    public void setCustomData15(String customData15) {
+        this.customData15 = customData15;
+    }
+
+    public String getCustomData15() {
+        return customData15;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData16
+     */
+    public void setCustomData16(String customData16) {
+        this.customData16 = customData16;
+    }
+
+    public String getCustomData16() {
+        return customData16;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData17
+     */
+    public void setCustomData17(String customData17) {
+        this.customData17 = customData17;
+    }
+
+    public String getCustomData17() {
+        return customData17;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData18
+     */
+    public void setCustomData18(String customData18) {
+        this.customData18 = customData18;
+    }
+
+    public String getCustomData18() {
+        return customData18;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData19
+     */
+    public void setCustomData19(String customData19) {
+        this.customData19 = customData19;
+    }
+
+    public String getCustomData19() {
+        return customData19;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData20
+     */
+    public void setCustomData20(String customData20) {
+        this.customData20 = customData20;
+    }
+
+    public String getCustomData20() {
+        return customData20;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData21
+     */
+    public void setCustomData21(String customData21) {
+        this.customData21 = customData21;
+    }
+
+    public String getCustomData21() {
+        return customData21;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData22
+     */
+    public void setCustomData22(String customData22) {
+        this.customData22 = customData22;
+    }
+
+    public String getCustomData22() {
+        return customData22;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData23
+     */
+    public void setCustomData23(String customData23) {
+        this.customData23 = customData23;
+    }
+
+    public String getCustomData23() {
+        return customData23;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData24
+     */
+    public void setCustomData24(String customData24) {
+        this.customData24 = customData24;
+    }
+
+    public String getCustomData24() {
+        return customData24;
+    }
+
+    /**
+     * Optional free-form data Not enabled by default Must be activated for your organization
+     *
+     * @param customData25
+     */
+    public void setCustomData25(String customData25) {
+        this.customData25 = customData25;
+    }
+
+    public String getCustomData25() {
+        return customData25;
     }
 
     public String getExperimentName() {
@@ -436,6 +725,24 @@ public class BitmovinAnalyticsConfig implements Parcelable {
                 this.getCustomData5(),
                 this.getCustomData6(),
                 this.getCustomData7(),
+                this.getCustomData8(),
+                this.getCustomData9(),
+                this.getCustomData10(),
+                this.getCustomData11(),
+                this.getCustomData12(),
+                this.getCustomData13(),
+                this.getCustomData14(),
+                this.getCustomData15(),
+                this.getCustomData16(),
+                this.getCustomData17(),
+                this.getCustomData18(),
+                this.getCustomData19(),
+                this.getCustomData20(),
+                this.getCustomData21(),
+                this.getCustomData22(),
+                this.getCustomData23(),
+                this.getCustomData24(),
+                this.getCustomData25(),
                 this.getExperimentName());
     }
 
@@ -447,6 +754,24 @@ public class BitmovinAnalyticsConfig implements Parcelable {
         this.setCustomData5(customData.getCustomData5());
         this.setCustomData6(customData.getCustomData6());
         this.setCustomData7(customData.getCustomData7());
+        this.setCustomData8(customData.getCustomData8());
+        this.setCustomData9(customData.getCustomData9());
+        this.setCustomData10(customData.getCustomData10());
+        this.setCustomData11(customData.getCustomData11());
+        this.setCustomData12(customData.getCustomData12());
+        this.setCustomData13(customData.getCustomData13());
+        this.setCustomData14(customData.getCustomData14());
+        this.setCustomData15(customData.getCustomData15());
+        this.setCustomData16(customData.getCustomData16());
+        this.setCustomData17(customData.getCustomData17());
+        this.setCustomData18(customData.getCustomData18());
+        this.setCustomData19(customData.getCustomData19());
+        this.setCustomData20(customData.getCustomData20());
+        this.setCustomData21(customData.getCustomData21());
+        this.setCustomData22(customData.getCustomData22());
+        this.setCustomData23(customData.getCustomData23());
+        this.setCustomData24(customData.getCustomData24());
+        this.setCustomData25(customData.getCustomData25());
         this.setExperimentName(customData.getExperimentName());
     }
 }